Extract plotting code
